Update: Github Actions + BuildMachineTest
Added new BuildMachineTest application to run on GithubActions. Will use this to slowly test this Renderer project on different os. Will make sure this builds and runs with Windows first. Then I will enable another job and see if that compiles and runs on other platforms.

This should work on Windows, but I will let Github decide that!