Update: DirectX12
I tried to get something up and running on GitHub actions for Windows/Linux using OpenGL. Nothing I tried worked though, the Windows action was passing but the Linux action kept failing. I couldn't work out what libraries to find and link.

SO.

I've now switch over to attempting to learn DirectX12 instead and try to understand how this is setup using this tutorial: https://www.3dgep.com/learning-directx-12-1/

This code displays a Window on screen and initializes everything that's needed for DX12 (command queue, swap chain, command list etc.). Trying to go through this tutorial but trying to work up the implementation to be abstracted as much as possible so I could hopefully one day implement Vulkan too.

The idea is that anything under the "Shared/" folder is considered a common base class and then anything under "DirectX12/" is specific to DirectX12 and these classes will inherit from classes under the Shared folder.

This seems to work locally so committing now!

Deleted previous implementations for OpenGL and any thirdparty folders associated with it. Removed the windows-test.yml workflow for now until I get confident enough with DirectX12, will be seeing if the build workflow passes with this commit.