Update: Smart Points + Scripts
Added smart points for core classes that are access from the Renderer.
Split EntryPoint up into separate macros and then passing through a time for any "AutoQuit" applications. Sandbox is a normal application I can use to test any changes and go through the tutorial I want! BuildMachineTest is an application that I want to only last for X amount of seconds, so I've implemented a macro to pass through an "AutoQuitTime" to accommodate this.